Optimize task sorting a bit & sort tags
The old __date_comp method was only called in 3 places and passed an
string parameter to determine what attributes to consider. However, the
caller already knows what attributes to get, so move it to the methods
itself to avoid useless branches.

__date_comp is "replaced" with __date_comp_continue, which basically
"continues" comparing the dates for further attributes, such as tags and
titles.

In the old code, the copy of a list of tags would get sorted and
compared for every compare. Since there is apparantly many comparisons,
when inserting tags, they'll also get sorted since their order doesn't
really seem to matter, avoiding the sorting overhead.
Additionally, it uses .get_tags_name_raw() to avoid the copy since it is
sorted anyway.

As an added bonus, instead of using locale.strxfrm and having to compare
2 strings twice (< and >), it now uses locale.strcoll, which basically
does the same thing, but is (probably) more optimized.

Note that it won't "reverse" when only the dates differ, since this isn't
done in the original code either. It also avoids creating that local
function, because it would just be used only once anyway and thus is
inlined.